+++ public/scripts/index.js
@@ -1,10 +1,3 @@
-
-const escapeFunc = function(str) {
-  let div = document.createElement("div");
-  div.appendChild(document.createTextNode(str));
-  return div.innerHTML;
-};
-
 const loadStories = function() {
   $.ajax({
     type: 'GET',
@@ -12,20 +5,12 @@ const loadStories = function() {
   }).then(function(data) {
     $('.stories').empty();
     console.log("We got the stories!")
-    // renderStories(data);
+    console.log("Data:", data);
   })
 };
 
-loadStories();
-
-// const renderStories = function(stories) {
-//   let story;
-//   for (let data in stories) {
-//     story = callback(tweets[data]);
-//     $(".stories").prepend(story);
-//   }
-// };
-
 $(document).ready(function() {
+  console.log("123");
   loadStories();
 });
+
