+++ server.js
@@ -47,7 +47,7 @@ app.use('/stories', storiesRoutes);
 app.use('/stories-api', storiesApiRoutes);
 app.use('/contributions', contributionsRoutes);
 app.use('/contributions-api', contributionsApiRoutes);
-
+//Each one needs own folder in db
 
 // Home page
 // Warning: avoid creating more routes in this file!
