+++ views/index.ejs
@@ -25,7 +25,7 @@
       <p>Collaborate with others to craft tales of swashbuckling adventure, fiery romance, or gut-busting comedy. The sky’s the limit.</p>
     </section>
     <section>
-      <h1 class="stories">Browse stories</h1>
+      <h1>Browse stories</h1>
       <!-- programmatically insert story preview divs here -->
     </section>
   </main>
